fix(gorsel): palet tek renge kilitleniyordu ve son kare tekrar ediyordu [DW-109] (#17)
* fix(gorsel): palet tek renge kilitleniyordu ve son kare tekrar ediyordu

Kullanici uc sikayet bildirdi: "bir resmi birden fazla kez kullanmissin",
"cok fazla ai resimler var sanki cartoon gibi", "resim ciktilari hep ayni
renkler tonlar". Ucu de olculdu, ucu de kod kusuru cikti — model degil.

PALET. Olculdu (2026-08-09, Mohenjo-Daro): 7 karenin hepsi 25°-47° ton
araliginda, dairesel ton yayilimi 0,007 (0 = tek renk). GORSEL_DIL "vary
lighting, time of day, weather" diyordu ama bu sahne ICINDE bir istek;
sahneler ayri isteklerle uretildigi icin modelin "onceki kare" bilgisi yok
ve her biri ayni "sicak tarih belgeseli" ortalamasina dusuyor. Cozum
kadrajdaki ile ayni (DW-107): dilek yerine her sahneye ACIKCA baska bir
isik. Ayni 5 sahne, yalnizca prompt degisti: yayilim 0,002 → 0,433, ton
araligi 25°-34° → 31°-202°.

CIZGI FILM. Sebep model degil promptun kendi istegiydi: GORSEL_DIL (b)
sikki "richly coloured historical painting or a detailed period
illustration" diyordu. Gecmisteki sahneler Commons'ta nadiren fotografla
karsilandigi icin AI dolgusunun cogu bu sikka dusuyor — cizim, kanalin
varsayilan gorunumu olmustu. Sikk muze kalitesinde bir CANLANDIRMA
fotografina cevrildi; "no invented event presented as a surviving
photograph" guvencesi yerinde birakildi, cunku fotogerceklik arttikca o
guvence daha da onemli.

TEKRAR. Olculdu: 7 sahne × 5 sn = 35,00 sn ama ses 35,88 sn. MPT acigi
itertools.cycle ile TAM bir klip ekleyerek kapatiyor, yani video sahne
1'in tekrariyla bitiyordu — kapanis cumlesi geri donusturulmus acilis
karesine dusuyordu. Son alti kosumun ucunde var. Klip suresi artik sabit
degil, ses/sahne. Sure TAHMIN EDILMIYOR, olculuyor: ayni ses ve hizda
kelime/saniye 2,05-2,79 arasinda degisiyor (tarih ve sayi yogun metin en
yavasi, bu kanalin metinleri tam oyle) ve %30 hata klip hesabini anlamsiz
kilardi. edge-tts yerel, ucretsiz ve deterministik — Mohenjo metni burada
35,88 sn olculdu, gercek kosumun sesi de 35,88 sn'ydi.

Klip suresi ondalik olabilmeli (35,88 / 7 = 5,13): tamsayi kisiti
tekrar ile son sahnenin dusmesi arasinda secim yapmaya zorluyordu.
KLIP_PAYI ikisini birden imkansiz kiliyor, test her iki yonu de
parametrik kilitliyor.

⚠️ KARE_DILI'nin isiktan bahseden iki maddesi kadraj tarifine cevrildi.
Birakilsalardi ayni sahnenin isigini iki kez, farkli sekilde tarif
ederlerdi — DW-107'de kapatilan "prompt kendisiyle celisiyor" tuzaginin
aynisi. Bir test bunu kilitliyor.

Okunabilirlik tabani: isik donusumu eklenince iki kare 0,21-0,22
parlakliga dustu. Shorts telefonda izleniyor; karanlik kare orada
okunmuyor. Prompt artik ana oznenin kucuk ekranda okunur kalmasini sart
kosuyor.

Notion: DW-109

* feat(olcum): ton yayilimi da kosumun yanina kaydediliyor

Palet kusurunu bulmak icin gecici bir betik yazmak gerekti; bir sonraki
gerilemeyi kimse fark etmezdi. DW-107'nin kurdugu desen: olcum kapi degil,
kosumun yanindaki kayit.

⚠️ Iki AYRI eksen kaydediliyor cunku biri iyilesirken digeri kotulesebilir.
Mohenjo-Daro kosumunda tam oyle oldu: yapisal tekrar sifirdi ama ton
yayilimi 0,007'ydi — tek sayiya bakmak 'duzeldi' yanilgisi verirdi.

Ton dairesel bir buyukluk (0° ile 359° komsu), duz ortalama yaniltir;
dairesel ortalama vektorunun uzunlugu kullaniliyor. Pikseller doygunlukla
agirliklandiriliyor ki gri alanlar rastgele tonlariyla olcumu bulandirmasin.

Notion: DW-109

* fix(arsiv): ayni gorsel iki sahnede kullanilamiyor

Kullanicinin "bir resmi birden fazla kez kullanmissin" sikayetinin IKINCI
sebebi. Ilki klip tekrariydi ve kapatildi; bu ayri bir yol.

Olculdu (2026-08-09, Library of Alexandria kosumu): sahne 3 ve sahne 7 AYNI
gravurdu — "The Great Library of Alexandria - Colorized.jpg" ile
"Ancientlibraryalex.jpg", algisal benzerlik 0,836. Ayni kosumda sahne 2 ve 5
de ayni sutunun (Pompey Sutunu) iki farkli fotografiydi, 0,734.

`used_titles` zaten vardi ve ikisini de yakalayamadi: dosya adlari tamamen
farkli. Benzerlik pikselden olculmeli, bu yuzden aday INDIRILDIKTEN sonra
bakiliyor.

Esik veriden secildi, tahminle degil: iki gercek tekrar 0,73 ve ustunde,
sonraki cift 0,58 — 0,70 aradaki bosluga dusuyor. Ayni oznenin baska
fotografi bilerek kapsama alindi; izleyici icin "ayni gorsel" ile "ayni
oznenin baska fotografi" arasinda fark yok.

⚠️ Kapi YUMUSAK: reddedilen adayin yerine sonraki aday deneniyor, hicbiri
kalmazsa yine de kabul ediliyor. Sahneyi bos birakmak ya AI dolgusu (para)
ya da eksik video demek olurdu. Bu tasarim esikteki belirsizligi de tasiyor:
veri tek kosumdan (n=1, temizlik oncekileri siliyor), esik yanlissa bedeli
bir arama daha.

Olcumler `gorsel_olcum` modulune tasindi. Sebep dairesel bagimlilik:
`wikimedia_materials` parmak izine ihtiyac duyuyor ama `youtube_automation`
zaten onu import ediyor. Iki taraf da ayni uygulamayi kullaniyor; bir test
kopyalanmadigini kilitliyor.

Notion: DW-109

+++ app/models/schema.py
@@ -82,7 +82,10 @@ class VideoParams(BaseModel):
     video_aspect: Optional[VideoAspect] = VideoAspect.portrait.value
     video_concat_mode: Optional[VideoConcatMode] = VideoConcatMode.random.value
     video_transition_mode: Optional[VideoTransitionMode] = None
-    video_clip_duration: Optional[int] = 5
+    # Ondalik: sahne basina sure sesin tam bolumu olmadiginda MPT ya bastan bir
+    # klibi tekrar ediyor ya da son sahneyi hic gostermiyor. Tamsayi kisiti
+    # ikisinden birini kacinilmaz kiliyordu (bkz. cli.py --video-clip-duration).
+    video_clip_duration: Optional[float] = 5
     video_clip_speed: Optional[float] = 1.0
     match_materials_to_script: bool = False
     video_count: Optional[int] = 1
